requirements.txt
mysql-connector-python>=8.0.0
flask>=2.3.0
